feat (map-tiles) : image ok & obstacle in progress
