+++ src/App.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import Grid from './components/Grid.vue'
+import Grid from './components/map/Grid.vue'
 
 export default {
   name: 'App',
